Introduce `flatMap` and `currentAndFlatMap`
This is a lot of code to hopefully handle complexity, that has two sources: (a) the way I store state in the URL and (b) the way I store the mapDataConfig in a const (and maybe later DB).

The URL needs a flat data structure (the way it is build ATM) and the mapDataConfig benefits from a hierarchical structure.

This code will hopefully be the glue in between.