MapboxStyle script: Remove visibility if present
This attribute tells us if the style is visible in Mapbox. However, this fact is not important to us, since we handle the visibility ourselves. We need to clean it so that the layers actually show up in our app, which does not handle this edge case (since it should not have to…)