SelectTopic: Use Dropdown (with Popper.js)
The Dropdown is cut of by the parent overflow-hidden. Headless UI has example code to use PopperJS with React Portal to Solve this. The "usePopper" stuff is not at the right place, yet; we need to move this once we add Popovers somewhere else or abstract the Dropdown in a component.