Introduce the concept of Legends
A Legend is a set of mapbox style layers. Each legend will get a icon next. And maybe we make them clickable to disable the given styles layers.

The style layers are also a helper to make sure the legend is complete (AKA all layers are described by a legend entry).