Extracting the OsmUrl did not work; lets keep it inline
This is not worth the hassle. We could keep the function in the Metadata file, but that is unexpected. Having it separate requires us to include it the main processing file even though it is only used indirectly there – very unexpected. Also we don't use it out of Metadata ATM. So lets just keep it inline.