roadClassification: Fix order of layers, bikelanes above roadClassification
This is how it works: The order is based on the URL param `config` which gets turned into `flattenConfigTopics` in <SourcesAndLayers/>. The param is base on the the `themes` config. So we change the order of themes and the order of topics in themes to change the implicit order in the map.