Metatag: Back to NextJS Metadata; Remove Helmet
`react-helmet` (NFL) did not work at all.

`react-helmet-async` (New York Times) worked a bit, but did not show the right page title and data. Some re-rendering race condition I assume.

So this gets us back to NextJS, which has a nice hierarchy of attributes to render metatags.

For this to work, all layouts and pages need to be Server components, which turned out to be easy on the n'th try.