Remove the getPublicRegion/s which caused confusion
Right now, all region data is public. So no need to split it up and make things confusing.

The bug was, that non-public regions could not be opened anymore which should be possible. The query was meant to make sure only public data of the region is exposed, however right now all data is public…