WIP
TEST: Use regular withDefaults (no migration for now); Move initialization out of wrapper component

Move config param migration step into parser

This does guard the application against undefined states where the initial URL is invalid. However, it also causes weird state effects.
1. Open page with broken config-param
2. App fixes the config-param and renders fine but does not fix the URL
3. Only once the user triggers a setConfig, the fixed URL is populated

Work around DB query by using staticData

This is not a great solution, because the static data will move to the DB at some point, but for now it is good enough and makes sure we do not query the DB.

FIX: Use same function signiture(?) for every usage with initialization present all the time

This is wasteful since we have to query the region every time…

TEST: Make changes easier to read