Subcategories: Refactor "npm run updateStyles", update code
The legacy styles are renamed and will not be updated anymore. We can migrate them step by step.

"npm run updateStyles" now creates separate `.ts`-files for each groups which we then import in the subcategories direclty.

This gives us a better visibility on changes, which pieces need to be updated in the code and which style groups are unused.