Introduce "npm run updateTilesets"
A new script that will download the given data for a given bbox (use a smaller fallback if that fails…), then run tippecanoe with some settings that allow us to style the data in mapbox studio in different zoom levels.

Afterwards the mbtiles have to be uploaded to mapbox manually.

Ping https://github.com/FixMyBerlin/private-issues/issues/1080