Maproulette API and Inspector Campaign-Hints
This introduces a way to create API endpoints that can be used in MapRoulette to create Projects/Campaigns to improve certain data categories. The geojson that is creates has a `task_markdown` field that can be used in MapRoulette as dynamic task description. The same text is reused in the Inspector to give a hint on what to do to update the data. The wording is a bit off there, because it is mainly meant to be used in Maproulette.

Ping https://github.com/FixMyBerlin/private-issues/issues/1364

Co-Authored-By: Laurenz Rasche <25564498+rush42@users.noreply.github.com>