Inspector: Rework tooltip/additional info for property values; Add new oneway values and tooltip
We use the details>summary element for those elments until now. But this requires a HTML structure that was quite complex.

This introduces a new component that styles a headless UI Disclosure and applies this component to all the places that used the details>summary pattern.

It also adds data on the reworked implicit oneway data.