Datasets: Fix updateStaticDataset Script for non-transformed data
I refactored the script recently to use the folder name as file name and not the original file name. That worked for transformed files, but not for untransformed files.

This makes sure we handle untransformed files the same way as transformed files, just with a different file name "*.transformed*" postfix.