Config param: Move shared utils out of v1 scope
Those methods are shared between v1 and v2, so we should move them to a shared scope to signal this shared usage.

This includes cleaning up functions that are not used anymore (they where only used when the encoding of v1 was still used) and renaming files to reflect their main function.