Export: Fix where missing keys caused a mixup of values
Our export functions forced all jsonb output to conform to the same structure. However, not all features have the same properties in the jsonb column. The nil cases resulted in a broken property representation of the data where eg. values from the meta column would become values of the tags. The solution is to always return an empty string so all tags are present but empty.