Parking: Apply capacity fallback to parking in obstacles table
This uses a different approach than the off_street_parking because our obstacles usually don't need this. However we miss-use this obstacle setup to also handle the street side parking data.
Which is why we have a flag now that is used to apply the capacity fallback based on area data.